Fix href attribute value extraction
