+++ sample/Pages/Shared/_Layout.cshtml
@@ -28,6 +28,30 @@
           </a>
 
           <a
+            href="/privacy"
+            class="group flex items-center px-3 py-2 text-sm font-medium rounded-md"
+            default-class="text-gray-600 hover:bg-gray-50 hover:text-gray-900"
+            current-class="bg-gray-200 text-gray-900"
+            aria-current-state="Page"
+          >
+            <heroicon-outline
+              icon="EyeOff"
+              class=" flex-shrink-0 -ml-1 mr-3 h-6 w-6"
+              default-class="text-gray-400 group-hover:text-gray-500"
+              current-class="text-gray-500"
+            />
+            <span class="truncate">Privacy - href</span>
+
+            <span
+              class="ml-auto inline-block py-0.5 px-3 text-xs rounded-full"
+              current-class="bg-gray-50"
+              default-class="bg-gray-200 text-gray-600 group-hover:bg-gray-200"
+            >
+              5
+            </span>
+          </a>
+
+          <a
             asp-area="" asp-page="/Privacy"
             class="group flex items-center px-3 py-2 text-sm font-medium rounded-md"
             default-class="text-gray-600 hover:bg-gray-50 hover:text-gray-900"
@@ -40,15 +64,7 @@
               default-class="text-gray-400 group-hover:text-gray-500"
               current-class="text-gray-500"
             />
-            <span class="truncate">Privacy</span>
-
-            <span
-              class="ml-auto inline-block py-0.5 px-3 text-xs rounded-full"
-              current-class="bg-gray-50"
-              default-class="bg-gray-200 text-gray-600 group-hover:bg-gray-200"
-            >
-              5
-            </span>
+            <span class="truncate">Privacy - asp-page</span>
           </a>
 
           <a
